edit directory structure and add gitignore and add start.sh
