+++ requirements.txt
@@ -1,3 +1,2 @@
-python-dotenv
-fastapi
-uvicorn
+fastapi == 0.79.0
+uvicorn == 0.18.2
