+++ src/Model/Login/LoginModel.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel
 
+
 class LoginModel(BaseModel):
     id : str
     pwd : str
